LFQBench散点图数据接口开发完毕

+++ src/main/java/net/csibio/propro/algorithm/lfqbench/bean/PeptideRatio.java
@@ -1,5 +1,5 @@
 package net.csibio.propro.algorithm.lfqbench.bean;
 
-public record PeptideRatio(String peptide, float x, float y) {
+public record PeptideRatio(String peptide, double x, double y) {
 
 }
